added login view and added restrict for users

+++ base/views.py
@@ -1,6 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 # rooms = [
     
@@ -10,22 +16,72 @@ from .forms import RoomForm
 # ]
 
 
+def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+        
+
+    
+    if request.method == 'POST':
+        username=request.POST.get('username')
+        password=request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+            
+        
+        user = authenticate(request,username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, 'Username or password does not exist')
+            
+        
+    
+    context ={}
+    return render(request,'base/login_register.html',context )
+    
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
-    rooms = Room.objects.filter(topic__name__contains=q)    
+    rooms = Room.objects.filter(
+        Q(topic__name__contains=q) |
+        Q(name__icontains=q)|
+        Q(description__icontains=q))    
+    
+    
     topics = Topic.objects.all()
+    room_count = rooms.count()
+    
+    
     context = {'rooms':rooms, 
-               'topics':topics}
+               'topics':topics,
+               'room_count':room_count}
     return render(request, 'base/home.html', context)
 
 
 def room(request,pk):
     room = Room.objects.get(pk=pk)
-            
     context = {'room':room}
     return render(request, 'base/room.html' ,context)
 
 
+
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -40,9 +96,14 @@ def createRoom(request):
 
 
 
+
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,8 +116,15 @@ def updateRoom(request,pk):
 
 
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id =pk)
+    
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
+    
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
